Move vacomplex::boundary(), star(), closure(), and opening() to algorithms.h (#992)
